feat: create standalone ModalContainer and ModalCloseButton components
