+++ packages/react/src/modal/index.js
@@ -1,9 +1,12 @@
 import Modal from './Modal';
 import ModalBody from './ModalBody';
+import ModalCloseButton from './ModalCloseButton';
+import ModalContainer from './ModalContainer';
 import ModalContent from './ModalContent';
 import ModalFooter from './ModalFooter';
 import ModalHeader from './ModalHeader';
 import ModalOverlay from './ModalOverlay';
+import useModal from './useModal';
 
 Modal.Body = ModalBody;
 Modal.Content = ModalContent;
@@ -14,8 +17,11 @@ Modal.Overlay = ModalOverlay;
 export {
   Modal,
   ModalBody,
+  ModalCloseButton, // internal use only
+  ModalContainer, // internal use only
   ModalContent,
   ModalFooter,
   ModalHeader,
   ModalOverlay,
+  useModal,
 };
